Apply changes requested in PR review
