+++ components/legalPage.tsx
@@ -1,0 +1,96 @@
+"use client";
+import React from "react";
+
+export default function LegalPage() {
+  return (
+    <main className="max-w-2xl mx-auto py-12 px-4">
+      <h1 className="text-3xl font-bold mb-8">Legal Notice</h1>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">Terms of Use</h2>
+        <p>
+          By accessing and using UniQ, you agree to comply with all applicable laws, regulations, and university guidelines. The content provided on this website is for informational and e-commerce purposes only and does not constitute legal advice.
+        </p>
+      </section>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">Copyright</h2>
+        <p>
+          All content, images, materials, and intellectual property on this site are owned by UniQ or its licensors. Unauthorized use, reproduction, or distribution of any material from this website is strictly prohibited.
+        </p>
+      </section>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">Disclaimer</h2>
+        <p>
+          UniQ is provided "as is," with no warranties of any kind, either express or implied. We make no representations regarding the accuracy, completeness, or reliability of the website's content. UniQ is not responsible for any direct, indirect, incidental, or consequential damages arising from the use or inability to use the website, including but not limited to data loss, business interruption, or any other issues.
+        </p>
+      </section>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">Limitation of Liability</h2>
+        <p>
+          UniQ’s liability is limited to the maximum extent permitted by law. We are not liable for any third-party services or external websites linked to or from UniQ, nor are we responsible for the privacy practices or content of those sites.
+        </p>
+      </section>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">Privacy Policy</h2>
+        <p>
+          By using UniQ, you consent to the collection, use, and storage of personal data as outlined in our Privacy Policy. We take your privacy seriously and will take all reasonable measures to protect your data in accordance with applicable privacy laws.
+        </p>
+      </section>
+      <hr className="my-8" />
+      <h1 className="text-3xl font-bold mb-8">Privacy Policy</h1>
+      <section className="mb-8">
+        <p>
+          At UniQ, we are committed to protecting your privacy and ensuring the security of your personal data. This Privacy Policy outlines the types of personal information we collect, how we use it, and the steps we take to safeguard your data.
+        </p>
+      </section>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">1. Information We Collect</h2>
+        <p>
+          When you use UniQ, we collect the following types of information:
+        </p>
+        <ul className="list-disc pl-6 mb-2">
+          <li><b>Personal Information:</b> When you make a purchase, contact us, or use certain features of the website, we may collect your name, email address, phone number, shipping address, payment information, and course details.</li>
+          <li><b>Usage Data:</b> We automatically collect information about your usage of the website, such as browsing history, search queries, and interactions with the site. This helps us improve your user experience.</li>
+        </ul>
+      </section>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">2. How We Use Your Information</h2>
+        <ul className="list-disc pl-6 mb-2">
+          <li>To process orders and provide you with the items you request.</li>
+          <li>To improve the functionality and user experience of the website.</li>
+          <li>To communicate with you about orders, promotions, or other relevant updates.</li>
+          <li>To analyze trends and user behavior to enhance the website's performance.</li>
+        </ul>
+      </section>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">3. Data Protection and Security</h2>
+        <p>
+          We take reasonable measures to protect your personal information from unauthorized access, alteration, or disclosure. All sensitive data, such as payment information, is encrypted using secure technologies. We also employ security protocols to protect your data when you interact with our site.
+        </p>
+      </section>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">4. Sharing Your Information</h2>
+        <p>
+          We do not sell or rent your personal information to third parties. However, we may share your information with trusted service providers who assist us in fulfilling orders, processing payments, or providing other services related to the website. These third parties are obligated to protect your data and use it only for the purposes specified by us.
+        </p>
+      </section>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">5. Cookies and Tracking Technologies</h2>
+        <p>
+          UniQ uses cookies and similar tracking technologies to improve website functionality, remember preferences, and analyze website usage. You can control cookie settings through your browser, but disabling cookies may affect some features of the website.
+        </p>
+      </section>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">6. Your Rights</h2>
+        <p>
+          You have the right to access, update, or delete your personal information at any time. If you wish to exercise these rights, or if you have any concerns about your data, please contact us at <a href="mailto:softeng310project@gmail.com" className="text-blue-600 underline">softeng310project@gmail.com</a>.
+        </p>
+      </section>
+      <section className="mb-8">
+        <h2 className="text-xl font-semibold mb-2">7. Changes to This Privacy Policy</h2>
+        <p>
+          We reserve the right to update this Privacy Policy from time to time. Any changes will be posted on this page with an updated revision date.
+        </p>
+      </section>
+    </main>
+  );
+}
